main/mcp-endpoint-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
configparser==6.0.0
asyncio-mqtt==0.16.1
pycryptodome==3.23.0
loguru==0.7.2